add current_user dependency

+++ routers/posts.py
@@ -9,6 +9,8 @@ import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+from auth import CurrentUser
+
 router = APIRouter()
 
 # Get Posts
@@ -25,17 +27,15 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
 
 # Create Post
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found!")
-
+async def create_post(
+    post: PostCreate, 
+    current_user: CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
+    
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id
+        user_id=current_user.id
     )
 
     db.add(new_post)
@@ -60,30 +60,24 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 
 # Update Post Full
 @router.put("/{post_id}", response_model=PostResponse)   # path parameter
-async def update_post_full(post_id: int,
-                    post_data: PostCreate,   # will contain data from request body when updating post
-                    db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(
+    post_id: int,
+    post_data: PostCreate,   # will contain data from request body when updating post
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
     results = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = results.scalars().first()
 
-    # if post exists
+    # Check if the post exists
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found!")
 
-    # if user exists
-    if post_data.user_id != post.user_id:
-        result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
+    # Prevent users from updating posts they don't own
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post!")
 
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found!"
-            )
-        
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -92,15 +86,21 @@ async def update_post_full(post_id: int,
 
 # Update Post Partial
 @router.patch("/{post_id}", response_model=PostResponse)   # path parameter
-async def update_post_partial(post_id: int,
-                    post_data: PostUpdate,   # will contain data from request body when updating post
-                    db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(
+    post_id: int,
+    post_data: PostUpdate,   # will contain data from request body when updating post
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
     results = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = results.scalars().first()
 
-    # if post exists
+    # Checks if the post exists
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found!")
+
+    # Prevent users from updating posts they don't own
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post!")
 
     # Pydantic method '(exclude_unset=True)' that converts a model into a dictionary, excluding fields that the client did not provide.
     update_data = post_data.model_dump(exclude_unset=True)
@@ -116,12 +116,19 @@ async def update_post_partial(post_id: int,
 
 # Delete Post
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)   # path parameter
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, 
+                    current_user: CurrentUser, 
+                    db: Annotated[AsyncSession, Depends(get_db)]):
     results = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = results.scalars().first()
 
+    # Checks if the post exists
     if not post:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found!")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found!")
+
+    # Prevent users from updating posts they don't own
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post!")
 
     await db.delete(post)
     await db.commit()
